app-code/requirements.txt
Flask==2.0.1
werkzeug==2.0.3
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
plotly==5.13.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
sqlalchemy==1.4.23
statsmodels==0.13.5
yfinance==0.1.63
pika==1.2.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.26.0
textblob==0.15.3
kafka-python==2.0.2
pyspark==3.2.0